requirements.txt
fastapi
uvicorn
playwright
google-cloud-storage
python-dotenv
